requirements.txt
dlib
matplotlib
numpy
Pillow
pytorch-msssim
scipy
torch
torchvision
torchsummary
tqdm
opencv-python
